add index.html with chatgpt prompt: build a static page with mandatory textinputs for spotname,latitude,longitude
checkboxed for 16 winddirections all are checked by default

sliders for minWindSpeed (3.0 to 25.0 default 12.0) maxWindSpeed (10.0 to 70.0 default 40.0), maxGustSpeedOntop (5.0 to 30.0 default 15.0) and hoursOfMatchingConditions (1 to 10 default 2)

an add button which adds the data to an array

each array item is represented by a chip which can be removed by clicking the x on it

below everything is a readonly textarea showing a link to the relative path wind.ics and the array items are get parameters to the url. Each input will result in a separate get parameter. The names of the parameters are repeated for each array item. By clicking on it the textarea the link is copied to clipboard and a copied hint is shown for 3 seconds.

use - as separator for winddirections

the lable on the chip schould be the name. the parameter selectedDirections should be named selected wind directions. the generated link should use calshow:// as url scheme

if a required field is missing add e red error message below the field

latitude,longitude should be validated to be a floatingpoint number